assets/controller.yaml: support using aws config for credentials
xref: https://issues.redhat.com/browse/CO-1298

The controller currently accepts the access key & secret key directly from the secret. This makes sure the pod starts after the secret is populated.

With this change, we mount the same secret to the pod to ensure that the pod starts after secret is created. But the env vars sources are updated to optional
since these keys might no longer be present in secret that is using aws config to configure the AWS SDK.

`AWS_SDK_LOAD_CONFIG` and `AWS_CONFIG_FILE` env variables are added to make sure the SDK tried to load the crendetials from the file specified. In cases
where the secret does not have the credentials key, this file will not exist and SDK will skip non-readable files see https://github.com/aws/aws-sdk-go/blob/87f0aa4be0153f21470e7a9802ebfb3403faa9ce/aws/session/shared_config.go#L164-L166
The aws-ebs-csi-driver already uses the default workflow for getting the session see https://github.com/openshift/aws-ebs-csi-driver/blob/master/pkg/cloud/cloud.go#L227 which should
support the AWS SDK config + STS.

What about when both access key & secret key and credential file is present in the secret. For cred-minter managed secrets these will have the same content, therefore it does not matter which the SDk uses, but for
user provided secret, the preference is decided by the AWS SDK (as it should be) which currently is env > config.

A bound-sa-token is added to the pod and mounted to the csi-driver pod to allow AWS SDK config file to point to it for use with AssumeRoleWithWebhookToken sts call.