support custom CA bundle for AWS API
When a cluster is installed in a AWS C2S region, access to the AWS
API requires a custom CA bundle for trust. The custom CA bundle
is read from the "ca-bundle.pem" key of the kube-cloud-config
ConfigMap in the openshift-config-managed namespace.

To use the custom CA bundle in the CSI driver pod, the operator
will sync the kube-cloud-config ConfigMap from the
openshift-config-managed namespace to the operator's namespace.

The CSI driver uses the CA bundle by having the AWS_CA_BUNDLE
environment variable point to the CA bundle file. Unfortunately,
the AWS SDK requires that, if the AWS_CA_BUNDLE environemnt variable
is set, then the file must exist. We cannot simply always set the
environment variable and rely on the presence or absence of the
ConfigMap to determine the behavior. Consequently, the CSI driver
deployment manifest has been templatized to fill out the parts
needed to support a custom CA bundle only when there is a custom CA
bundle.

https://issues.redhat.com/browse/CORS-1584