dynamically update controller asset for custom CA bundle
The determination about whether a custom CA bundle is being used
is assessed when the operator starts. This can lead to incorrect
behavior if the kube-cloud-config ConfigMap has not yet been created
in the openshit-config-managed namespace when the operator starts.
Instead of only looking for the ConfigMap at start-up, these changes
will have the operator looking for the ConfigMap on every reconcile
and adjusting the controller deployment accordingly.